Started implementing the dictionary.
I opted to implement it using typescript files instead
of a json file, so I can make the entire dictionary
a typesafe monster that can work with anything.

I'm forcing it to be a server-only thing so this will ensure
that nextjs cache everything in the server side so the
app will be performant even though it's a ts file

At first I build the lang-switch now using even imported icons
for it. and passing the files from server component to a
client component. This is the power of typescript dictionary.